% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{action_scan}
\alias{action_scan}
\title{Scan en cours}
\usage{
action_scan(session, values, scan_id)
}
\arguments{
\item{session}{Session shiny}

\item{values}{Valeurs réactives}

\item{scan_id}{ID du scan}
}
\value{
TRUE
}
\description{
Scan en cours
}
