% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{actu_AI}
\alias{actu_AI}
\title{Extraire les infos actuelles des IA}
\usage{
actu_AI(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos actuelles des IA
}
