% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{actu_enigmes}
\alias{actu_enigmes}
\title{Extraire les infos actuelles des énigmes}
\usage{
actu_enigmes(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos actuelles des énigmes
}
