% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{actu_scans}
\alias{actu_scans}
\title{Extraire les infos actuelles des scan}
\usage{
actu_scans(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos actuelles des scan
}
