% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{appendChatMessage}
\alias{appendChatMessage}
\title{Ajouter un message dans le chat}
\usage{
appendChatMessage(sender, message, ns)
}
\arguments{
\item{sender}{Expéditeur}

\item{message}{Message}

\item{ns}{id shiny}
}
\value{
JS
}
\description{
Ajouter un message dans le chat
}
