% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{check_scan}
\alias{check_scan}
\title{Check si le scan est déjà fait}
\usage{
check_scan(values, scan_id)
}
\arguments{
\item{values}{Valeurs réactives}

\item{scan_id}{ID du scan}
}
\value{
boolean
}
\description{
Check si le scan est déjà fait
}
