% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{EcranAdminServer}
\alias{EcranAdminServer}
\title{Serveur Admin}
\usage{
EcranAdminServer(id, values)
}
\arguments{
\item{id}{id}

\item{values}{valeurs réactives}
}
\value{
shiny server
}
\description{
Serveur Admin
}
