% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{EcranAIServer}
\alias{EcranAIServer}
\title{Serveur de l'IA}
\usage{
EcranAIServer(id, values)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}
}
\value{
shiny server
}
\description{
Serveur de l'IA
}
