% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{EcranControleServer}
\alias{EcranControleServer}
\title{Serveur Contrôle}
\usage{
EcranControleServer(id, base_url)
}
\arguments{
\item{id}{id}

\item{base_url}{base de l'url pour les écrans}
}
\value{
shiny server
}
\description{
Serveur Contrôle
}
