% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{EcranQRServer}
\alias{EcranQRServer}
\title{Serveur Questions/Réponses}
\usage{
EcranQRServer(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny server
}
\description{
Serveur Questions/Réponses
}
