% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{EcranQRUI}
\alias{EcranQRUI}
\title{UI Questions/Réponses}
\usage{
EcranQRUI(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny ui
}
\description{
UI Questions/Réponses
}
