% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{EcranScanServer}
\alias{EcranScanServer}
\title{Serveur de scans}
\usage{
EcranScanServer(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny server
}
\description{
Serveur de scans
}
