% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{EcranScanUI}
\alias{EcranScanUI}
\title{UI scans}
\usage{
EcranScanUI(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny ui
}
\description{
UI scans
}
