% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{EcranSourceServer}
\alias{EcranSourceServer}
\title{Serveur du code source}
\usage{
EcranSourceServer(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny server
}
\description{
Serveur du code source
}
