% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{EcranSourceUI}
\alias{EcranSourceUI}
\title{UI code source}
\usage{
EcranSourceUI(id, values, local)
}
\arguments{
\item{id}{id}

\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
shiny ui
}
\description{
UI code source
}
