% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeRoom-package.R
\docType{package}
\name{escapeRoom-package}
\alias{escapeRoom}
\alias{escapeRoom-package}
\title{escapeRoom: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdelc/escapeRoom}
  \item Report bugs at \url{https://github.com/tdelc/escapeRoom/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
