% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{i_question}
\alias{i_question}
\title{Extraire l'id de la question en cours}
\usage{
i_question(values, local = TRUE)
}
\arguments{
\item{values}{Valeurs réactives}

\item{local}{Valeurs locales}
}
\value{
number
}
\description{
Extraire l'id de la question en cours
}
