% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{indicAnimServer}
\alias{indicAnimServer}
\title{Module serveur animation}
\usage{
indicAnimServer(id, db_i, path)
}
\arguments{
\item{id}{id du module}

\item{db_i}{db des indices}

\item{path}{chemin des fichiers}
}
\value{
moduleServer
}
\description{
Module serveur animation
}
