% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{indicTextServer}
\alias{indicTextServer}
\title{Module serveur text}
\usage{
indicTextServer(id, db_i)
}
\arguments{
\item{id}{id du module}

\item{db_i}{db des indices}
}
\value{
moduleServer
}
\description{
Module serveur text
}
