% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{info_AI}
\alias{info_AI}
\title{Extraire les infos initiales des IA}
\usage{
info_AI(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos initiales des IA
}
