% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{info_enigmes}
\alias{info_enigmes}
\title{Extraire les infos initiales des enigmes}
\usage{
info_enigmes(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos initiales des enigmes
}
