% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{info_scans}
\alias{info_scans}
\title{Extraire les infos initiales des scans}
\usage{
info_scans(values)
}
\arguments{
\item{values}{Valeurs réactives}
}
\value{
data.frame
}
\description{
Extraire les infos initiales des scans
}
