% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{LabelServer}
\alias{LabelServer}
\title{Module serveur label}
\usage{
LabelServer(id, db, path)
}
\arguments{
\item{id}{id du module}

\item{db}{db question ou indice}

\item{path}{chemin des fichiers}
}
\value{
moduleServer
}
\description{
Module serveur label
}
