% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_db_enigmes}
\alias{load_db_enigmes}
\title{Chargement de la DB enigmes}
\usage{
load_db_enigmes(id_drive)
}
\arguments{
\item{id_drive}{id du google drive}
}
\value{
data.frame
}
\description{
Chargement de la DB enigmes
}
