% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_db_scans}
\alias{load_db_scans}
\title{Chargement de la DB scans}
\usage{
load_db_scans(id_drive)
}
\arguments{
\item{id_drive}{id du google drive}
}
\value{
data.frame
}
\description{
Chargement de la DB scans
}
