% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_db_trad}
\alias{load_db_trad}
\title{Chargement de la DB traduction}
\usage{
load_db_trad(id_drive)
}
\arguments{
\item{id_drive}{id du google drive}
}
\value{
data.frame
}
\description{
Chargement de la DB traduction
}
