% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{new_prompt}
\alias{new_prompt}
\title{Changement du prompt de l'IA}
\usage{
new_prompt(new_priority, nb_mails = 5000, language = "fr")
}
\arguments{
\item{new_priority}{Nouvelles priorité}

\item{nb_mails}{Nombre de mails envoyés}
}
\value{
text
}
\description{
Changement du prompt de l'IA
}
