% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_list}
\alias{style_list}
\title{Style HTML par couleur}
\usage{
style_list(color_dark, color_light)
}
\arguments{
\item{color_dark}{Couleur foncée}

\item{color_light}{Couleur claire}
}
\value{
html
}
\description{
Style HTML par couleur
}
