% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_scan}
\alias{style_scan}
\title{Style HTML pour les scans}
\usage{
style_scan(color_dark, color_light)
}
\arguments{
\item{color_dark}{Couleur foncée}

\item{color_light}{Couleur claire}
}
\value{
html
}
\description{
Style HTML pour les scans
}
