% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temps_restant}
\alias{temps_restant}
\title{Calcul de temps restant en seconde à partir d'une heure et minute}
\usage{
temps_restant(heure, minute)
}
\arguments{
\item{heure}{(en numérique)}

\item{minute}{(en minute)}
}
\value{
liste avec le total des secondes
}
\description{
Calcul de temps restant en seconde à partir d'une heure et minute
}
\examples{
temps_restant(23,30)
}
