% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{titre_scan}
\alias{titre_scan}
\title{Extraite titre du Scan}
\usage{
titre_scan(values, scan_id)
}
\arguments{
\item{values}{Valeurs réactives}

\item{scan_id}{ID du scan}

\item{session}{Session shiny}
}
\value{
string
}
\description{
Extraite titre du Scan
}
