% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{trad}
\alias{trad}
\title{Fonction de traduction}
\usage{
trad(text, values)
}
\arguments{
\item{text}{texte à traduire, provenant du fr}

\item{values}{valeurs réactives}
}
\value{
text
}
\description{
Fonction de traduction
}
